+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Notable changes of tdmclient. Release versions refer to [https://pypi.org/projec
 ### Added
 
 - Functions `get_client()` and `get_node()` added to repl
-- Functions `get_nodes()` and `list()` added to notebook
+- In notebooks:
+	- functions `get_nodes()` and `list()` to get the list of nodes available in the tdm
+	- decorator `@sync_var` to synchronize the robot variables accessed in the function
 
 ### Fixed